le am pus in pachet si test la strategy

+++ src/main/java/org/example/model/ValueStock.java
@@ -1,14 +1,14 @@
 package org.example.model;
 
 public class ValueStock extends Stock{
-    private int dividend_rate;
+    private double dividend_rate;
 
-    public ValueStock(int id, String name, int price, int company, int market, int dividend_rate) {
+    public ValueStock(int id, String name, int price, int company, int market, double dividend_rate) {
         super(id, name, price, company, market);
         this.dividend_rate = dividend_rate;
     }
 
-    public int getDividend_rate() {
+    public double getDividend_rate() {
         return dividend_rate;
     }
